requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
appdirs=1.4.4=py_0
argon2-cffi=20.1.0=py38h7b6447c_1
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
brotlipy=0.7.0=py38h8df0ef7_1001
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py38h578d9bd_1
cffi=1.14.4=py38h261ae71_0
chardet=4.0.0=py38h578d9bd_1
click=7.1.2=pyh9f0ad1d_0
cryptography=3.2.1=py38h7699a38_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py_2
dataclasses=0.7=pyhb2cacf7_7
datasets=1.2.1=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.3=pypi_0
entrypoints=0.3=py38_0
expat=2.2.10=he6710b0_2
filelock=3.0.12=pyh9f0ad1d_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
gensim=3.8.3=pypi_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=py38h32f6830_1
jupyter_core=4.6.3=py38_0
jupyter_highlight_selected_word=0.2.0=py38h578d9bd_1002
jupyter_latex_envs=1.4.6=py38h32f6830_1001
jupyter_nbextensions_configurator=0.4.1=py38h578d9bd_2
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38h82cb98a_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libxslt=1.1.34=hc22bd24_0
lxml=4.6.1=py38hbb43d70_0
lz4-c=1.9.3=h2531618_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.1.2=py38h4fdacc2_0
matplotlib-base=3.3.3=py38h5c7f4ab_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
multiprocess=0.70.11.1=pypi_0
mypy_extensions=0.4.3=py38_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
ninja=1.10.2=py38hff7bd54_0
nltk=3.5=pypi_0
notebook=6.1.4=py38_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1h=h516909a_0
packaging=20.8=pyhd3deb0d_0
pandas=1.2.1=py38ha9443f7_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.1.0=py38he98fc37_0
pip=20.3.3=py38h06a4308_0
pot=0.7.0=py38h950e882_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
ptyprocess=0.6.0=py38_0
pyarrow=3.0.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h578d9bd_3
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2020.5=pyhd3eb1b0_0
pyyaml=5.3.1=py38h8df0ef7_1
pyzmq=19.0.2=py38he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
regex=2020.11.13=py38h25fe258_0
requests=2.25.1=pyhd3deb0d_0
sacremoses=0.0.43=pyh9f0ad1d_0
scikit-learn=0.24.1=pypi_0
scipy=1.5.2=py38h0b6359f_0
send2trash=1.5.0=py38_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
smart-open=4.1.2=pypi_0
sqlite=3.33.0=h62c20be_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.4=py38h781e378_1
toml=0.10.1=py_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.1=py38h25fe258_0
tqdm=4.49.0=pypi_0
traitlets=5.0.5=py_0
transformers=4.2.2=pyhd8ed1ab_0
typed-ast=1.4.2=py38h27cfd23_1
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
xxhash=2.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
